Fixes lack of privileges for global router deletion
Commit 89b70c5deee68afcc72a06d900ab87a18108df99
provided a fix for defect DE794. Unfortately that
fix only worked for admin users. For non-admin
users the bug remains. This patch fixes that
problem by ensuring context is elevated in the
required places in the code so that the operations
can succeed.

This patch corresponds to commit:
1db31f6731fe45236acde80bb2d7fcde4b926313 in neutron
branch asr1k_kilo_stable_wip.

Fixes defects: DE804