ASR1K plugin: error condition handling
This is a porting of DE260 from Junoplus.
Also reported as DE777 on kilo/liberty based:
https://rally1.rallydev.com/#/26206778171d/detail/defect/44092650098

The issue is clearing the router gateway causes cisco-cfg-agent to
get stuck trying to already removed config from the ASR and causing
continuous traceback in cisco-cfg-agent.log file.

Hence this fix is to catch the exception caused by the ASR router
on delete/remove operations (ex: in this case it happened due to
REMOVE_SUBINTERFACE). This will make that the message gets removed
from the queue (which will free up netconf from getting stuck
retrying the failed operation that caused the Exception).