Fixes REST API URL path for cfg agent scheduler
The URL path for the hosting device to cfg agent scheduler was incorrect.
This patch fixes that. The patch also updates the attributes dict
for the routertype, hosting_device and hosting_device_template resources
so that the required attributes are correct.

Finally, the patch adds a new REST API method, get_hosting_device_config,
for the hosting device resource. This method returns the running config
of a hosting device.

Closes defects: DE875, DE877

Change-Id: I68495982de86d63bfb9a028d00af446dc9112dbd