Fix defect DE903 (Fail to clear default route)
The cfg agent fails when clearing default route for the tenant routers when the external gateway is cleared.

Root cause: This happens only when the tenant router is the last tenant router with an external gateway. Thus when it is cleared the global router (owning the external interface) is also removed.  But the 1C code modified code caused the global routers to be deleted before the tenant router is processed. As part of the global router delete, the external subinterface is removed. But default route for the tenant router was based on this and when the cfg agent comes to this part, it fails.

We have now fixed the order of this processing, so that removed router are processed after updaed routers are processed. Also global routers are processed last in either way.

Change-Id: If76edd8ff774c9d59da1098c0f607b2ad426fa59