Add RPC to allow config agent to update hosting device status
This patch adds a RPC callback so that the config agent can
request the plugin to update status for a hosting device.
The patch also updates the router schedulers so that it takes
into account hosting device status when it does scheduling.
A hosting device will only be considered a candidate if its
status is Active.

Change-Id: I7b30644154e51279532902b4a98b523e8e535ec3