Monitor status of config agents more comprehensively
This patch improves the status monitoring of config agents.
The status of config agent is, like before, checked when a
notification is to be sent for a hosting device. But the
status is now also checked if no notification has been sent
to a config agent for a certain time interval. This interval
can be set using a new configuration option
'cfg_agent_monitoring_interval'.
Additional UTs will be provided in a subsequent patch.

Change-Id: Ie634bf6bd82cda9560e866726133d9db78be2e40