Handle multiple concurrent scheduling attempts for the same router
This patch handles the IntegrityError exception that can happen
when multiple API processes simultaneously try to schedule an
unscheduled router. When this exception is encountered by a process
it just needs to abort the scheduling since another process has
taken care of it.

The patch also breaks out some common code in the schedule/un-schedule
processing into its own function.

Closes bug: DE1073

Change-Id: I07f767201729b3108f2aad40394556f50eb6eebb